Move launcher/__init__.py into launch. run_tests.py runs unbuffered.
