+++ launch
@@ -1,17 +1,240 @@
 #!/usr/bin/env python3
 
 import argparse
+import atexit
+import itertools
 import os
 import os.path
 import pexpect
-import subprocess
+import random
+import shlex
+import shutil
 import signal
+import subprocess
 import sys
-
-from launcher import *
+import time
 
 
 TMUX_CONF = './.tmux.conf'
+MIN_PORT = 24000
+MAX_PORT = 32000
+
+
+def RandomPort():
+    portlock = None
+    ports = list(range(MIN_PORT, MAX_PORT))
+    while not portlock and ports:
+        port = random.choice(ports)
+        ports.remove(port)
+        portlock = f'/tmp/launch-port-{port}.lock'
+        try:
+            fd = os.open(portlock, os.O_CREAT | os.O_EXCL | os.O_CLOEXEC)
+            os.close(fd)
+
+            def unlock():
+                try:
+                    os.remove(portlock)
+                except OSError:
+                    pass
+
+            atexit.register(unlock)
+            return port
+        except OSError:
+            portlock = None
+    raise RuntimeError('Run out of TCP port locks!')
+
+
+CONFIG = {
+    'board': 'malta',
+    'config': {
+        'debug': False,
+        'gdbport': RandomPort(),
+        'graphics': False,
+        'elf': 'sys/mimiker.elf',
+        'initrd': 'initrd.cpio',
+        'args': [],
+        'board': {
+            'malta': {
+                'kernel': 'sys/mimiker.elf',
+            },
+            'rpi3': {
+                'kernel': 'sys/mimiker.img.gz',
+            },
+        },
+    },
+    'qemu': {
+        'options': [
+            '-nodefaults',
+            # Configure record/replay function for deterministic replay,
+            # refer to https://github.com/qemu/qemu/blob/master/docs/replay.txt
+            # '-icount', 'shift=3,sleep=on,rr=record,rrfile=replay.bin',
+            '-kernel', '{kernel}',
+            '-initrd', '{initrd}',
+            '-S', '-gdb', 'tcp:127.0.0.1:{gdbport},server,wait',
+            '-serial', 'none'],
+        'board': {
+            'malta': {
+                'binary': 'qemu-mimiker-mipsel',
+                'options': [
+                    '-device', 'VGA',
+                    '-device', 'rtl8139',
+                    '-machine', 'malta',
+                    '-cpu', '24Kf'],
+                'uarts': [
+                    dict(name='/dev/tty1', port=RandomPort(), raw=True),
+                    dict(name='/dev/tty2', port=RandomPort()),
+                    dict(name='/dev/cons', port=RandomPort())
+                ]
+            },
+            'rpi3': {
+                'binary': 'qemu-mimiker-aarch64',
+                'options': [
+                    '-machine', 'raspi3',
+                    '-smp', '4',
+                    '-cpu', 'cortex-a53'],
+                'uarts': [
+                    dict(name='/dev/cons', port=RandomPort())
+                ]
+            }
+        }
+    },
+    'gdb': {
+        'pre-options': [
+            '-n',
+            '-ex=set confirm no',
+            '-iex=set auto-load safe-path {}/'.format(os.getcwd()),
+            '-ex=set tcp connect-timeout 30',
+            '-ex=target remote localhost:{gdbport}',
+            '--silent',
+        ],
+        'extra-options': [],
+        'post-options': [
+            '-ex=source .gdbinit',
+            '-ex=continue',
+        ],
+        'board': {
+            'malta': {
+                'binary': 'mipsel-mimiker-elf-gdb'
+            },
+            'rpi3': {
+                'binary': 'aarch64-mimiker-elf-gdb'
+            }
+        }
+    }
+}
+
+
+def setboard(name):
+    setvar('board', name)
+    # go over top-level configuration variables
+    for top, config in CONFIG.items():
+        if type(config) is not dict:
+            continue
+        board = getvar('board.' + name, start=config, failok=True)
+        if not board:
+            continue
+        # merge board variables into generic set
+        for key, value in board.items():
+            if key not in config:
+                config[key] = value
+            elif type(value) == list:
+                config[key].extend(value)
+            else:
+                raise RuntimeError(f'Cannot merge {top}.board.{name}.{key} '
+                                   f'into {top}')
+        # finish merging by removing alternative configurations
+        del config['board']
+
+
+def getvar(name, start=CONFIG, failok=False):
+    value = start
+    for f in name.split('.'):
+        try:
+            value = value[f]
+        except KeyError as ex:
+            if failok:
+                return None
+            raise ex
+    return value
+
+
+def setvar(name, val, config=CONFIG):
+    fs = name.split('.')
+    while len(fs) > 1:
+        config = config[fs.pop(0)]
+    config[fs.pop(0)] = val
+
+
+def getopts(*names):
+    opts = itertools.chain.from_iterable([getvar(name) for name in names])
+    return [opt.format(**getvar('config')) for opt in opts]
+
+
+class Launchable():
+    def __init__(self, name, cmd):
+        self.name = name
+        self.cmd = cmd
+        self.window = None
+        self.pid = None
+        self.options = []
+
+    def start(self, session):
+        cmd = ' '.join([self.cmd] + list(map(shlex.quote, self.options)))
+        self.window = session.new_window(
+            attach=False, window_name=self.name, window_shell=cmd)
+        self.pid = int(self.window.attached_pane._info['pane_pid'])
+
+    def spawn(self):
+        return pexpect.spawn(self.cmd, self.options)
+
+    def stop(self):
+        if self.pid is None:
+            return
+
+        time.sleep(0.2)
+        try:
+            os.kill(self.pid, signal.SIGKILL)
+        except ProcessLookupError:
+            # Process has already quit!
+            pass
+        self.pid = None
+
+
+class QEMU(Launchable):
+    def __init__(self):
+        super().__init__('qemu', getvar('qemu.binary'))
+
+        self.options = getopts('qemu.options')
+        for uart in getvar('qemu.uarts'):
+            port = uart['port']
+            self.options += ['-serial', f'tcp:127.0.0.1:{port},server,wait']
+
+        if getvar('config.args'):
+            self.options += ['-append', ' '.join(getvar('config.args'))]
+        if getvar('config.debug'):
+            self.options += ['-S']
+        if not getvar('config.graphics'):
+            self.options += ['-display', 'none']
+
+
+class GDB(Launchable):
+    def __init__(self):
+        super().__init__('gdb', getvar('gdb.binary'))
+
+        self.options += getopts(
+                'gdb.pre-options', 'gdb.extra-options', 'gdb.post-options')
+        self.options.append(getvar('config.elf'))
+
+
+class SOCAT(Launchable):
+    def __init__(self, name, tcp_port, raw=False):
+        super().__init__(name, 'socat')
+        # The simulator will only open the server after some time has
+        # passed.  To minimize the delay, keep reconnecting until success.
+        stdio_opt = 'STDIO'
+        if raw:
+            stdio_opt += ',cfmakeraw'
+        self.options = [stdio_opt, f'tcp:localhost:{tcp_port},retry,forever']
 
 
 def TestRun(sim, dbg, timeout):
@@ -80,8 +303,6 @@ def DevelRun(sim, dbg):
 
 
 if __name__ == '__main__':
-    setup_terminal()
-
     parser = argparse.ArgumentParser(
         description='Launch kernel in Malta board simulator.')
     parser.add_argument('args', metavar='ARGS', type=str,
